chore(SDK-4042): renames method param, moves method
